ifcfg-rh: write zero-value bridge and bridge port options
Zero values are actually valid values for various bridge options
and should be written out.  Otherwise, when reading the property
back in, it gets assigned the default value which is often not
zero, causing the wrong value to be set in the connection.

Only properties with default values should not be written out.