core: ignore gateway host route during connection generation
When a gateway is not in the prefix of any of the interface's IP addresses,
NetworkManager adds a static host route to the gateway through the
interface to ensure the gateway can be reached.  That route will not
be part of the persistent connection (since it was added automatically)
but would normally be picked up by connection generation.  This would
cause the generated connection not to match with the persistent
connection, because the persistent connection does not have the host
route.  Ignore the gateway host route when capturing the interface's
existing IP configuration.